+++ resource/csdk/connectivity/common/inc/uarraylist.h
@@ -39,8 +39,8 @@ extern "C"
 typedef struct u_arraylist_t
 {
     void **data;
-    uint32_t length;
-    uint32_t capacity;
+    size_t length;
+    size_t capacity;
 } u_arraylist_t;
 
 /**
@@ -86,7 +86,7 @@ void u_arraylist_shrink_to_fit(u_arraylist_t *list);
  * @param[in] index        index of array list.
  * @return void pointer of data if success or NULL pointer otherwise.
  */
-void *u_arraylist_get(const u_arraylist_t *list, uint32_t index);
+void *u_arraylist_get(const u_arraylist_t *list, size_t index);
 
 /**
  * Returns the index of the data from the array list.
@@ -95,7 +95,7 @@ void *u_arraylist_get(const u_arraylist_t *list, uint32_t index);
  * @param[out]index        index of array list.
  * @return true if success, false otherwise.
  */
-bool u_arraylist_get_index(const u_arraylist_t *list, const void *data,  uint32_t *index);
+bool u_arraylist_get_index(const u_arraylist_t *list, const void *data, size_t *index);
 
 /**
  * Add data in the array list.
@@ -111,14 +111,14 @@ bool u_arraylist_add(u_arraylist_t *list, void *data);
  * @param[in] index      index of array list.
  * @return void pointer of the data if success or NULL pointer otherwise.
  */
-void *u_arraylist_remove(u_arraylist_t *list, uint32_t index);
+void *u_arraylist_remove(u_arraylist_t *list, size_t index);
 
 /**
  * Returns the length of the array list.
  * @param[in] list       pointer of array list.
  * @return length of the array list.
  */
-uint32_t u_arraylist_length(const u_arraylist_t *list);
+size_t u_arraylist_length(const u_arraylist_t *list);
 
 /**
  * Returns whether the data exists or not.
